Store json on server

+++ src/main/java/eu/hansolo/ControllerV1.java
@@ -43,29 +43,33 @@ public class ControllerV1 {
                 try(FileOutputStream fos = new FileOutputStream(file.getFilename())) {
                     fos.write(file.getBytes());
                     File           jarfile  = new File("/app/" + file.getFilename());
-                    TreeNode<Item> treeNode = Scanner.getClassesAndMethods(jarfile.getName());
-                    String         jsonText = Helper.toJsonString(jarfile.getName(), treeNode, "", "", "");
-                    jarfile.delete();
 
-                    HttpResponse response = HttpResponse.ok(jsonText).contentType(MediaType.APPLICATION_JSON_TYPE).status(HttpStatus.OK);
-                    return response;
+                    TreeNode<Item> treeNode = Scanner.getClassesAndMethods("/app/" + jarfile.getName());
+                    if (null == treeNode) {
+                        return HttpResponse.badRequest("{}")
+                                               .contentType(MediaType.APPLICATION_JSON_TYPE)
+                                               .status(HttpStatus.BAD_REQUEST);
+                    } else {
+                        final String filename = jarfile.getName().replace(".jar", ".json");
+                        final String json     = Helper.toJsonString(jarfile.getAbsolutePath(), treeNode, "", "", "");
+                        Helper.saveTextFile(json, filename);
+                        jarfile.delete();
+                        return HttpResponse.ok(json).contentType(MediaType.APPLICATION_JSON_TYPE).status(HttpStatus.OK);
+                    }
                 } catch (Exception e) {
-                    HttpResponse response = HttpResponse.badRequest("{}")
-                                                        .contentType(MediaType.APPLICATION_JSON_TYPE)
-                                                        .status(HttpStatus.BAD_REQUEST);
-                    return response;
+                    return HttpResponse.badRequest("{}")
+                                       .contentType(MediaType.APPLICATION_JSON_TYPE)
+                                       .status(HttpStatus.BAD_REQUEST);
                 }
             } else {
-                HttpResponse response = HttpResponse.badRequest("{}")
-                                                    .contentType(MediaType.APPLICATION_JSON_TYPE)
-                                                    .status(HttpStatus.BAD_REQUEST);
-                return response;
+                return HttpResponse.badRequest("{}")
+                                   .contentType(MediaType.APPLICATION_JSON_TYPE)
+                                   .status(HttpStatus.BAD_REQUEST);
             }
         } catch (Exception e) {
-            HttpResponse response = HttpResponse.badRequest("{}")
-                                                .contentType(MediaType.APPLICATION_JSON_TYPE)
-                                                .status(HttpStatus.BAD_REQUEST);
-            return response;
+            return HttpResponse.badRequest("{}")
+                               .contentType(MediaType.APPLICATION_JSON_TYPE)
+                               .status(HttpStatus.BAD_REQUEST);
         }
     }
 }
